Move get_payment_model to keep backward compatibility
